test(e2e): Port cloudflare-workers to span streaming and keep a static copy
Renames cloudflare-workers to cloudflare-workers-static, which stays on
traceLifecycle: 'static' with the transaction-based specs, so the static
trace lifecycle keeps E2E coverage. cloudflare-workers-streaming becomes
the new cloudflare-workers and drops its explicit 'stream' pin, so the app
exercises the default lifecycle.
