+++ dev-packages/e2e-tests/test-applications/cloudflare-workers-static/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "cloudflare-workers-streaming",
+  "name": "cloudflare-workers-static",
   "version": "0.0.0",
   "private": true,
   "scripts": {
