+++ dev-packages/e2e-tests/test-applications/cloudflare-workers-static/src/index.ts
@@ -76,11 +76,11 @@ class MyDurableObjectBase extends DurableObject<Env> {
 
 export const MyDurableObject = Sentry.instrumentDurableObjectWithSentry(
   (env: Env) => ({
+    traceLifecycle: 'static',
     dsn: env.E2E_TEST_DSN,
     environment: 'qa', // dynamic sampling bias to keep transactions
     tunnel: `http://localhost:3031/`, // proxy server
     tracesSampleRate: 1.0,
-    traceLifecycle: 'stream',
     transportOptions: {
       // We are doing a lot of events at once in this test
       bufferSize: 1000,
@@ -91,11 +91,11 @@ export const MyDurableObject = Sentry.instrumentDurableObjectWithSentry(
 
 export default Sentry.withSentry(
   (env: Env) => ({
+    traceLifecycle: 'static',
     dsn: env.E2E_TEST_DSN,
     environment: 'qa', // dynamic sampling bias to keep transactions
     tunnel: `http://localhost:3031/`, // proxy server
     tracesSampleRate: 1.0,
-    traceLifecycle: 'stream',
     transportOptions: {
       // We are doing a lot of events at once in this test
       bufferSize: 1000,
