+++ dev-packages/e2e-tests/test-applications/cloudflare-workers-static/tests/tsconfig.json
@@ -1,8 +1,5 @@
 {
   "extends": "../tsconfig.json",
-  "compilerOptions": {
-    "types": ["@cloudflare/vitest-pool-workers"]
-  },
   "include": ["./**/*.ts"],
   "exclude": []
 }
