+++ dev-packages/e2e-tests/test-applications/cloudflare-workers-static/tsconfig.json
