+++ dev-packages/e2e-tests/test-applications/cloudflare-workers-static/wrangler.toml
@@ -1,5 +1,5 @@
 #:schema node_modules/wrangler/config-schema.json
-name = "cloudflare-workers-streaming"
+name = "cloudflare-workers-static"
 main = "src/index.ts"
 compatibility_date = "2024-07-25"
 compatibility_flags = ["nodejs_compat"]
