+++ dev-packages/e2e-tests/test-applications/cloudflare-workers/tests/tsconfig.json
@@ -1,5 +1,8 @@
 {
   "extends": "../tsconfig.json",
+  "compilerOptions": {
+    "types": ["@cloudflare/vitest-pool-workers"]
+  },
   "include": ["./**/*.ts"],
   "exclude": []
 }
